requirements2.txt
aiofiles
aiohttp
aiosignal
aliyun-python-sdk-core
aliyun-python-sdk-kms
altair
annotated-types
antlr4-python3-runtime
anyio
asttokens
async-timeout
attrs
Automat
beartype
blessed
buildtools
# causal-conv1d
certifi
cffi
chardet
charset-normalizer
clean-fid
click
# clip
cmake
colorama
coloredlogs
constantly
contourpy
crcmod
cryptography
cycler
decorator
decord
diffusers
docopt
easydict
einops
exceptiongroup
executing
fairscale
fastapi
ffmpeg
ffmpy
filelock
flatbuffers
fonttools
frozenlist
fsspec
ftfy
furl
gpustat
gradio
gradio_client
greenlet
h11
httpcore
httpx
huggingface-hub
humanfriendly
hyperlink
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
incremental
ipdb
ipython
jedi
Jinja2
jmespath
joblib
jsonschema
jsonschema-specifications
kiwisolver
kornia
lazy_loader
lightning-utilities
lit
lpips
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
# motion-vector-extractor
mpmath
multidict
mypy-extensions
networkx
ninja
# numpy
nvidia-cublas-cu11
nvidia-cublas-cu12
nvidia-cuda-cupti-cu11
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu11
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu11
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu11
nvidia-cudnn-cu12
nvidia-cufft-cu11
nvidia-cufft-cu12
nvidia-curand-cu11
nvidia-curand-cu12
nvidia-cusolver-cu11
nvidia-cusolver-cu12
nvidia-cusparse-cu11
nvidia-cusparse-cu12
nvidia-ml-py
nvidia-nccl-cu11
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu11
nvidia-nvtx-cu12
omegaconf
onnxruntime
open-clip-torch
opencv-python-headless==4.9.0.80
orderedmultidict
orjson
oss2
# packaging
pandas
parso
pexpect
pillow
piq
pkgconfig
prompt-toolkit
protobuf
psutil
ptflops
ptyprocess
pure-eval
pycparser
pycryptodome
pydantic
pydantic_core
pydub
Pygments
pynvml
pyparsing
pyre-extensions
python-dateutil
python-multipart
pytorch-lightning
pytz
PyYAML
redo
referencing
regex
requests
rich
rotary-embedding-torch
rpds-py
ruff
safetensors
scikit-image
scikit-learn
scipy
semantic-version
sentencepiece
shellingham
simplejson
six
sk-video
sniffio
SQLAlchemy
stack-data
starlette
sympy
thop
threadpoolctl
tifffile
timm
tokenizers
tomli
tomlkit
toolz
# torch
# torchaudio
torchdiffeq
torchmetrics
torchsde
# torchvision
tqdm
traitlets
trampoline
transformers
triton
Twisted
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
websockets
xformers==0.0.25.post1
yarl
zipp
zope.interface
onnxruntime-gpu
